- arch aware kvm wrappers
Some users might just want to call "kvm" without thinking if this needs
qemu-system-* nor which arch they are on nor to add --enable-kvm.

This feature already exists for x86, but there is no reason not to
extend that at least to the most commonly used targets (arm, s390x,
ppc).

Furthermore we can pull in some sanity checks e.g. to disable smt on
ppc64 to be even more usable.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>